raynet_saas_optimization/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pandas
python-dotenv
requests
pydantic==1.10.10
scikit-learn
xgboost
httpx
fastapi-limiter
redis

# pip install -r requirements.txt
# pip freeze   - control code


